move usage-tracking stuff to own module
