+++ src/wormhole_transit_relay/server_tap.py
@@ -10,7 +10,7 @@ from twisted.internet import protocol
 from autobahn.twisted.websocket import WebSocketServerFactory
 
 from . import transit_server
-from .server_state import create_usage_tracker
+from .usage import create_usage_tracker
 from .increase_rlimits import increase_rlimits
 from .database import get_db
